refactor: data.js
