+++ src/keyboard__row1.js
@@ -1,4 +1,4 @@
-import { arr } from '../data.js';
+import { arr } from './data.js';
 
 let key = 'en';
