+++ src/keyboard__row5.js
@@ -1,4 +1,4 @@
-import { arr } from '../data.js';
+import { arr } from './data.js';
 
 let key = 'en';
