+++ src/keyboard__rows.js
@@ -1,4 +1,4 @@
-import { arr } from '../data.js';
+import { arr } from './data.js';
 
 let key = 'en';
